Use radius to query in findSites instead of bbox
bbox queries from WQP have not been working for several months. This workaround allows users to query WQP sites in a similar manner as the bbox, but uses the center of the map view area and the distance to the corner of the map (rounded up 10%) to query, then trims sites to the bbox lat/longs.